add socket response to all client due to waiting service in clients

+++ sql/waiting/create.sql
@@ -4,7 +4,7 @@ create table waiting(
 	pid integer not null references patients(pid) unique,
 	vid integer null references visits(vid) unique,
     paper_id integer not null,
-	priority numeric not null,
+	priority integer not null,
     wait_start_time timestamp with time zone not null default current_timestamp,
 	unique(did,pid)
 )
